Risolto overflow dello score e dei nemici uccisi

+++ src/enemy/SkeletonEnemy.h
@@ -10,7 +10,7 @@ public:
     float       get_attack_distance() const override { return 0.9f; }
     uint8_t     get_base_damage()     const override { return 5; }
     std::string get_texture_name()    const override { return "skeleton"; }
-	uint8_t     get_reward_score()    const override { return 10; }
+	uint16_t     get_reward_score()    const override { return 10; }
     void        init_states()               override;
     void        on_target_in_range()        override;
     void        on_target_out_of_range()    override;
